Move both clients to root module
